refactor file logging backend
