+++ src/rust/logger/src/format.rs
@@ -1,3 +1,7 @@
+// Copyright 2021 Twitter, Inc.
+// Licensed under the Apache License, Version 2.0
+// http://www.apache.org/licenses/LICENSE-2.0
+
 use crate::Record;
 use rustcommon_time::{DateTime, Local, SecondsFormat};
